+++ lending/RELEASES.md
@@ -19,3 +19,13 @@ Based on:
 - [python v0.1.1] lending
 ### Releases
 - [PyPI v0.1.1] https://pypi.org/project/codat-lending/0.1.1 - lending
+
+## 2023-09-05 00:15:50
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.78.3 (2.96.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.1.2] lending
+### Releases
+- [PyPI v0.1.2] https://pypi.org/project/codat-lending/0.1.2 - lending